Clean up test structure, update error response messaging to include status code and text
